플레이어 애니메이션 추가 및 여러가지 수정
건든 것 목록:
Player.cs
그 밖에 애니메이션 쪽 조금

Enum.cs 추가: 앞으로 enum 여기에

+++ The_Nightmare/Component/MoveComponent.cs
@@ -8,10 +8,10 @@ namespace The_Nightmare
 {
     public class MoveComponent
     {
-        public void MoveBy(GameObject owner, int deltaX, int deltaY, int[,] map)
+        public void MoveBy(GameObject owner, double deltaX, double deltaY, int[,] map)
         {
-            int nextX = owner.X + deltaX;
-            int nextY = owner.Y + deltaY;
+            int nextX = (int)(owner.X + deltaX);
+            int nextY = (int)(owner.Y + deltaY);
 
             if(nextX >= 0 && nextX < map.GetLength(0) && nextY >= 0 && nextY < map.GetLength(1))
             {
